+++ Deploy/bash/deploy-s3-terraform-backend.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
 set -e # Turn on exit on error
-set -x # Turn on verbosity
+set +x # <-- Do not change this value!
+       # Logging is turned on in a later step based on CLOUD_TAMER_PARAM.
+       # CLOUD_TAMER_PARAM = false (Jenkins assumed; verbose logging turned off)
+       # CLOUD_TAMER_PARAM = true (Dev machine assumed; verbose logging turned on)
+       # NOTE: Setting the CLOUD_TAMER_PARAM to a value that does not match the
+       #       assumed host machine will cause the script to fail.
 
 #
 # Change to working directory (if not set by a parent script)
@@ -41,23 +46,41 @@ export AWS_ACCOUNT_NUMBER="${AWS_ACCOUNT_NUMBER_PARAM}"
 if [ "${CLOUD_TAMER_PARAM}" != "false" ] && [ "${CLOUD_TAMER_PARAM}" != "true" ]; then
   echo "ERROR: CLOUD_TAMER_PARAM parameter must be true or false"
   exit 1
-else
+elif [ "${CLOUD_TAMER_PARAM}" == "false" ]; then
+
+  # Turn off verbose logging for Jenkins jobs
+  set +x
+  echo "Don't print commands and their arguments as they are executed."
   CLOUD_TAMER="${CLOUD_TAMER_PARAM}"
+
+  # Import the "get temporary AWS credentials via AWS STS assume role" function
+  source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_aws_sts_assume_role.sh"
+
+else # [ "${CLOUD_TAMER_PARAM}" == "true" ]
+
+  # Turn on verbose logging for development machine testing
+  set -x
+  echo "Print commands and their arguments as they are executed."
+  CLOUD_TAMER="${CLOUD_TAMER_PARAM}"
+
+  # Import the "get temporary AWS credentials via CloudTamer API" function
+  source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_cloudtamer_api.sh"
+
 fi
 
 #
-# Define functions
+# Set AWS target environment
 #
 
-# Import the "get temporary AWS credentials via CloudTamer API" function
+if [ "${CLOUD_TAMER}" == "true" ]; then
+  fn_get_temporary_aws_credentials_via_cloudtamer_api "${AWS_ACCOUNT_NUMBER}" "${CMS_ENV}"
+else
+  fn_get_temporary_aws_credentials_via_aws_sts_assume_role "${AWS_ACCOUNT_NUMBER}" "${CMS_ENV}"
+fi
 
-# shellcheck source=./functions/fn_get_temporary_aws_credentials_via_cloudtamer_api.sh
-source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_cloudtamer_api.sh"
-
-# Import the "get temporary AWS credentials via AWS STS assume role" function
-
-# shellcheck source=./functions/fn_get_temporary_aws_credentials_via_aws_sts_assume_role.sh
-source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_aws_sts_assume_role.sh"
+#
+# Define function to create or verify dynamodb table for module
+#
 
 fn_create_or_verify_dynamodb_table_for_module ()
 {
@@ -79,19 +102,10 @@ fn_create_or_verify_dynamodb_table_for_module ()
       --attribute-definitions AttributeName=LockID,AttributeType=S \
       --key-schema AttributeName=LockID,KeyType=HASH \
       --provisioned-throughput ReadCapacityUnits=5,WriteCapacityUnits=5 \
-      1> /dev/null
+      1> /dev/null \
+      2> /dev/null
   fi
 }
-
-#
-# Set AWS target environment
-#
-
-if [ "${CLOUD_TAMER}" == "true" ]; then
-  fn_get_temporary_aws_credentials_via_cloudtamer_api "${AWS_ACCOUNT_NUMBER}" "${CMS_ENV}"
-else
-  fn_get_temporary_aws_credentials_via_aws_sts_assume_role "${AWS_ACCOUNT_NUMBER}" "${CMS_ENV}"
-fi
 
 #
 # Configure terraform
@@ -99,10 +113,20 @@ fi
 
 # Reset logging
 
-echo "Setting terraform debug level to $DEBUG_LEVEL..."
-export TF_LOG=$DEBUG_LEVEL
-export TF_LOG_PATH=/var/log/terraform/tf.log
-rm -f /var/log/terraform/tf.log
+if [ "${CLOUD_TAMER}" == "true" ]; then
+
+  # Enable terraform logging on development machine
+  echo "Setting terraform debug level to $DEBUG_LEVEL..."
+  export TF_LOG=$DEBUG_LEVEL
+  export TF_LOG_PATH=/var/log/terraform/tf.log
+  rm -f /var/log/terraform/tf.log
+
+else
+
+  # Disable terraform logging on Jenkins
+  export TF_LOG=
+
+fi
 
 #
 # Create or verify backend components
@@ -169,6 +193,7 @@ export S3_SERVER_ACCESS_LOGS_BUCKET="${GET_S3_SERVER_ACCESS_LOGS_BUCKET}"
 set +e # Turn off exit on error
 aws --region "${AWS_DEFAULT_REGION}" s3api get-bucket-policy \
   --bucket "${S3_SERVER_ACCESS_LOGS_BUCKET}" \
+  1> /dev/null \
   2> /dev/null
 BUCKET_POLICY_STATUS=$?
 set -e # Turn on exit on error
@@ -192,7 +217,9 @@ if [ "${LOG_DELIVERY_GROUP_PERMISSION_COUNT}" -lt 2 ]; then
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-acl \
     --bucket "${S3_SERVER_ACCESS_LOGS_BUCKET}" \
     --grant-write URI=http://acs.amazonaws.com/groups/s3/LogDelivery \
-    --grant-read-acp URI=http://acs.amazonaws.com/groups/s3/LogDelivery    
+    --grant-read-acp URI=http://acs.amazonaws.com/groups/s3/LogDelivery \
+    1> /dev/null \
+    2> /dev/null
 fi
 
 # Create or verify S3 tfstate bucket
@@ -217,6 +244,7 @@ export S3_TFSTATE_BUCKET="${GET_S3_TFSTATE_BUCKET}"
 set +e # Turn off exit on error
 aws --region "${AWS_DEFAULT_REGION}" s3api get-bucket-policy \
   --bucket "${S3_TFSTATE_BUCKET}" \
+  1> /dev/null \
   2> /dev/null
 BUCKET_POLICY_STATUS=$?
 set -e # Turn on exit on error
@@ -224,7 +252,9 @@ set -e # Turn on exit on error
 if [ "${BUCKET_POLICY_STATUS}" != "0" ]; then
   aws --region "${AWS_DEFAULT_REGION}" s3api put-public-access-block \
     --bucket "${S3_TFSTATE_BUCKET}" \
-    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true
+    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true \
+    1> /dev/null \
+    2> /dev/null
 fi
 
 # Add logging to the S3 tfstate bucket
@@ -239,8 +269,10 @@ if [ -z "${S3_TFSTATE_BUCKET_LOGGING}" ]; then
   j2 tfstate_bucket_logging.json.j2 -o generated/tfstate_bucket_logging.json
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-logging \
-  --bucket "${S3_TFSTATE_BUCKET}" \
-  --bucket-logging-status file://generated/tfstate_bucket_logging.json
+    --bucket "${S3_TFSTATE_BUCKET}" \
+    --bucket-logging-status file://generated/tfstate_bucket_logging.json \
+    1> /dev/null \
+    2> /dev/null
 fi
 
 # Add or verify versioning on the S3 tfstate bucket
@@ -250,7 +282,9 @@ S3_TFSTATE_BUCKET_VERSIONING=$(aws --region "${AWS_DEFAULT_REGION}" s3api get-bu
 
 if [ -z "${S3_TFSTATE_BUCKET_VERSIONING}" ]; then
   aws s3api put-bucket-versioning --bucket "${S3_TFSTATE_BUCKET}" \
-    --versioning-configuration Status=Enabled
+    --versioning-configuration Status=Enabled \
+    1> /dev/null \
+    2> /dev/null
 fi
 
 # Add or verify bucket policy on the S3 tfstate bucket
@@ -258,6 +292,7 @@ fi
 set +e # Turn off exit on error
 aws --region "${AWS_DEFAULT_REGION}" s3api get-bucket-policy \
   --bucket "${S3_TFSTATE_BUCKET}" \
+  1> /dev/null \
   2> /dev/null
 BUCKET_POLICY_STATUS=$?
 set -e # Turn on exit on error
@@ -270,7 +305,9 @@ if [ "${BUCKET_POLICY_STATUS}" != "0" ]; then
     
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-policy \
       --bucket "${S3_TFSTATE_BUCKET}" \
-      --policy file://generated/tfstate_bucket_policy.json
+      --policy file://generated/tfstate_bucket_policy.json \
+      1> /dev/null \
+      2> /dev/null
 fi
 
 # Add or verify server side encryption on the S3 tfstate bucket
@@ -278,6 +315,7 @@ fi
 set +e # Turn off exit on error
 aws --region "${AWS_DEFAULT_REGION}" s3api get-bucket-encryption \
   --bucket "${S3_TFSTATE_BUCKET}" \
+  1> /dev/null \
   2> /dev/null
 BUCKET_ENCRYPTION_STATUS=$?
 set -e # Turn on exit on error
@@ -290,7 +328,9 @@ if [ "${BUCKET_ENCRYPTION_STATUS}" != "0" ]; then
     
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-encryption \
     --bucket "${S3_TFSTATE_BUCKET}" \
-    --server-side-encryption-configuration file://generated/tfstate_bucket_server_side_encryption.json
+    --server-side-encryption-configuration file://generated/tfstate_bucket_server_side_encryption.json \
+    1> /dev/null \
+    2> /dev/null
 fi
 
 # Create or verify dynamodb table for core module
