+++ Deploy/bash/initialize-environment.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
-set -e #Exit on first error
-set -x #Be verbose
+set -e # Turn on exit on error
+set +x # <-- Do not change this value!
+       # Logging is turned on in a later step based on CLOUD_TAMER_PARAM.
+       # CLOUD_TAMER_PARAM = false (Jenkins assumed; verbose logging turned off)
+       # CLOUD_TAMER_PARAM = true (Dev machine assumed; verbose logging turned on)
+       # NOTE: Setting the CLOUD_TAMER_PARAM to a value that does not match the
+       #       assumed host machine will cause the script to fail.
 
 #
 # Change to working directory
@@ -35,8 +40,35 @@ fi
 if [ "${CLOUD_TAMER_PARAM}" != "false" ] && [ "${CLOUD_TAMER_PARAM}" != "true" ]; then
   echo "ERROR: CLOUD_TAMER_PARAM parameter must be true or false"
   exit 1
-else
+elif [ "${CLOUD_TAMER_PARAM}" == "false" ]; then
+
+  # Turn off verbose logging for Jenkins jobs
+  set +x
+  echo "Don't print commands and their arguments as they are executed."
   CLOUD_TAMER="${CLOUD_TAMER_PARAM}"
+
+  # Import the "get temporary AWS credentials via AWS STS assume role" function
+  source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_aws_sts_assume_role.sh"
+
+else # [ "${CLOUD_TAMER_PARAM}" == "true" ]
+
+  # Turn on verbose logging for development machine testing
+
+  set -x
+  echo "Print commands and their arguments as they are executed."
+  CLOUD_TAMER="${CLOUD_TAMER_PARAM}"
+
+  # Reset logging
+
+  echo "Setting terraform debug level to $DEBUG_LEVEL..."
+  export TF_LOG=$DEBUG_LEVEL
+  export TF_LOG_PATH=/var/log/terraform/tf.log
+  rm -f /var/log/terraform/tf.log
+
+  # Import the "get temporary AWS credentials via CloudTamer API" function
+
+  source "${START_DIR}/functions/fn_get_temporary_aws_credentials_via_cloudtamer_api.sh"
+
 fi
 
 #
@@ -97,12 +129,18 @@ S3_AUTOMATION_BUCKET_EXISTS=$(aws --region "${AWS_DEFAULT_REGION}" s3api list-bu
   --output text)
 
 if [ -z "${S3_AUTOMATION_BUCKET_EXISTS}" ]; then
+
   aws --region "${AWS_DEFAULT_REGION}" s3api create-bucket \
-    --bucket "${S3_AUTOMATION_BUCKET}"
+    --bucket "${S3_AUTOMATION_BUCKET}" \
+    1> /dev/null \
+    2> /dev/null
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-public-access-block \
     --bucket "${S3_AUTOMATION_BUCKET}" \
-    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true
+    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true \
+    1> /dev/null \
+    2> /dev/null
+
 fi
 
 # Create or verify S3 environment bucket
@@ -112,12 +150,18 @@ S3_ENVIRONMENT_BUCKET_EXISTS=$(aws --region "${AWS_DEFAULT_REGION}" s3api list-b
   --output text)
 
 if [ -z "${S3_ENVIRONMENT_BUCKET_EXISTS}" ]; then
+
   aws --region "${AWS_DEFAULT_REGION}" s3api create-bucket \
-    --bucket "${TARGET_CMS_ENV}"
+    --bucket "${TARGET_CMS_ENV}" \
+    1> /dev/null \
+    2> /dev/null
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-public-access-block \
     --bucket "${TARGET_CMS_ENV}" \
-    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true
+    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true \
+    1> /dev/null \
+    2> /dev/null
+
 fi
 
 #
@@ -133,20 +177,26 @@ if [ -z "${S3_CLOUDTRAIL_BUCKET_EXISTS}" ]; then
   # Create cloudtrail bucket
 
   aws --region "${AWS_DEFAULT_REGION}" s3api create-bucket \
-    --bucket "${TARGET_CMS_ENV}-cloudtrail"
+    --bucket "${TARGET_CMS_ENV}-cloudtrail" \
+    1> /dev/null \
+    2> /dev/null
 
   # Block public access
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-public-access-block \
     --bucket "${TARGET_CMS_ENV}-cloudtrail" \
-    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true
+    --public-access-block-configuration BlockPublicAcls=true,IgnorePublicAcls=true,BlockPublicPolicy=true,RestrictPublicBuckets=true \
+    1> /dev/null \
+    2> /dev/null
 
   # Give "Write objects" and "Read bucket permissions" to the "S3 log delivery group" of the "cloudtrail" bucket
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-acl \
     --bucket "${TARGET_CMS_ENV}-cloudtrail" \
     --grant-write URI=http://acs.amazonaws.com/groups/s3/LogDelivery \
-    --grant-read-acp URI=http://acs.amazonaws.com/groups/s3/LogDelivery
+    --grant-read-acp URI=http://acs.amazonaws.com/groups/s3/LogDelivery \
+    1> /dev/null \
+    2> /dev/null
 
   # Add bucket policy to the "cloudtrail" S3 bucket
 
@@ -155,16 +205,28 @@ if [ -z "${S3_CLOUDTRAIL_BUCKET_EXISTS}" ]; then
 
   aws --region "${AWS_DEFAULT_REGION}" s3api put-bucket-policy \
     --bucket "${TARGET_CMS_ENV}-cloudtrail" \
-    --policy file://ab2d-cloudtrail-bucket-policy.json
+    --policy file://ab2d-cloudtrail-bucket-policy.json \
+    1> /dev/null \
+    2> /dev/null
 
 fi
 
 # Reset logging
 
-echo "Setting terraform debug level to $DEBUG_LEVEL..."
-export TF_LOG=$DEBUG_LEVEL
-export TF_LOG_PATH=/var/log/terraform/tf.log
-rm -f /var/log/terraform/tf.log
+if [ "${CLOUD_TAMER}" == "true" ]; then
+
+  # Enable terraform logging on development machine
+  echo "Setting terraform debug level to $DEBUG_LEVEL..."
+  export TF_LOG=$DEBUG_LEVEL
+  export TF_LOG_PATH=/var/log/terraform/tf.log
+  rm -f /var/log/terraform/tf.log
+
+else
+
+  # Disable terraform logging on Jenkins
+  export TF_LOG=
+
+fi
 
 # Initialize and validate terraform for the target environment
 
@@ -192,12 +254,23 @@ terraform validate
 cd "${START_DIR}/.."
 cd "terraform/environments/${TARGET_CMS_ENV}"
 
-terraform apply \
-  --var "env=${TARGET_CMS_ENV}" \
-  --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
-  --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
-  --target module.iam \
-  --auto-approve
+if [ "${CLOUD_TAMER_PARAM}" == "true" ]; then
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --target module.iam \
+    --auto-approve
+else
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --target module.iam \
+    --auto-approve \
+    1> /dev/null \
+    2> /dev/null
+fi
 
 #
 # Create or verify KMS components
@@ -206,11 +279,21 @@ terraform apply \
 cd "${START_DIR}/.."
 cd "terraform/environments/${TARGET_CMS_ENV}"
 
-terraform apply \
-  --var "env=${TARGET_CMS_ENV}" \
-  --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
-  --target module.kms \
-  --auto-approve
+if [ "${CLOUD_TAMER_PARAM}" == "true" ]; then
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --target module.kms \
+    --auto-approve
+else
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --target module.kms \
+    --auto-approve \
+    1> /dev/null \
+    2> /dev/null
+fi
 
 #
 # Get secrets
@@ -260,14 +343,27 @@ fi
 cd "${START_DIR}/.."
 cd "terraform/environments/${TARGET_CMS_ENV}"
 
-terraform apply \
-  --var "env=${TARGET_CMS_ENV}" \
-  --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
-  --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
-  --var "ab2d_bfd_insights_s3_bucket=${AB2D_BFD_INSIGHTS_S3_BUCKET}" \
-  --var "ab2d_bfd_kms_arn=${AB2D_BFD_KMS_ARN}" \
-  --target module.iam_bfd_insights \
-  --auto-approve
+if [ "${CLOUD_TAMER_PARAM}" == "true" ]; then
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --var "ab2d_bfd_insights_s3_bucket=${AB2D_BFD_INSIGHTS_S3_BUCKET}" \
+    --var "ab2d_bfd_kms_arn=${AB2D_BFD_KMS_ARN}" \
+    --target module.iam_bfd_insights \
+    --auto-approve
+else
+  terraform apply \
+    --var "env=${TARGET_CMS_ENV}" \
+    --var "mgmt_aws_account_number=${CMS_ECR_REPO_ENV_AWS_ACCOUNT_NUMBER}" \
+    --var "aws_account_number=${TARGET_AWS_ACCOUNT_NUMBER}" \
+    --var "ab2d_bfd_insights_s3_bucket=${AB2D_BFD_INSIGHTS_S3_BUCKET}" \
+    --var "ab2d_bfd_kms_arn=${AB2D_BFD_KMS_ARN}" \
+    --target module.iam_bfd_insights \
+    --auto-approve \
+    1> /dev/null \
+    2> /dev/null
+fi
 
 #
 # Configure networking
@@ -292,5 +388,7 @@ if [ "${VPC_ENABLE_DNS_HOSTNAMES}" == "False" ]; then
   echo "Enabling DNS hostnames on VPC..."
   aws ec2 modify-vpc-attribute \
     --vpc-id "${VPC_ID}" \
-    --enable-dns-hostnames
+    --enable-dns-hostnames \
+    1> /dev/null \
+    2> /dev/null
 fi
